# src/main/resources/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bank Rest",
    "description": "Bank Rest API",
    "contact": {
      "name": "Dmitriy Folomkin",
      "url": "www.folomkin.dev",
      "email": "dmitriy.folomkin@yandex.ru"
    },
    "license": {
      "name": "License of API",
      "url": "API license URL"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8189",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
...
  "paths": {
...
    "/api/v1/users/change-balance": {
      "post": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Перевод средств между своими картами. Для Пользователя",
        "description": "Укажите последние 4 цифры номера карты-отправителя, карты-получателя и сумму перевода",
        "operationId": "changeCardBalance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardBalanceChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardBalanceChangeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{userid}": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Получение пользователя по id. Для Администратора",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "userid",
            "in": "path",
            "description": "Id пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/my-cards": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Получение списка своих карт пользователем. Для Пользователя",
        "operationId": "getMyCards",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "Номер страницы с результатом",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Количество выводимых карт на странице. Минимум 1, максимум 50",
            "required": false,
            "schema": {
              "maximum": 50,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Поле сортировки",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cardNumber",
            "in": "query",
            "description": "Укажите 4 последние цифры номера своей карты",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelCardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/filter-users": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Получение списка пользователей постранично. Для Администратора",
        "operationId": "getUsersPages",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "Номер страницы с результатом",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Количество выводимых пользователей на странице. Минимум 1, максимум 30",
            "required": false,
            "schema": {
              "maximum": 30,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Поле сортировки",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/block-request/{cardNumber}": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Отправка пользователем запроса на блокировку карты. Для Пользователя",
        "description": "Укажите последние 4 цифры номера",
        "operationId": "sendingBlockingRequest",
        "parameters": [
          {
            "name": "cardNumber",
            "in": "path",
            "description": "Последние 4 цифры номера карты",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/all": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Получение списка всех пользователей. Для Администратора",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          }
        }
      },
...
      "CardResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "encryptedNumber": {
            "type": "string",
            "description": "Номер карты с маской"
          },
          "expirationDate": {
            "type": "string",
            "description": "Дата окончания срока действия карты",
            "format": "date"
          },
          "cardStatus": {
            "type": "string",
            "description": "Статус карты",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "ISEXPIRATEDDATE",
              "BLOCKREQUEST"
            ]
          },
          "balance": {
            "type": "number",
            "description": "Баланс карты"
          },
          "owner": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "description": "Возвращаемый объект при регистрации карты"
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "description": "Пользователь карты"
      },
      "CardBalanceChangeRequest": {
        "type": "object",
        "properties": {
          "cardNumberOut": {
            "type": "string",
            "description": "Номер карты отправителя. Указать последние 4 цифры",
            "example": "1111"
          },
          "cardNumberIn": {
            "type": "string",
            "description": "Номер карты получателя. Указать последние 4 цифры",
            "example": "1111"
          },
          "amount": {
            "type": "number",
            "description": "Сумма перевода",
            "example": 12.3
          }
        },
        "description": "Форма перевода средств между картами"
      },
      "CardBalanceChangeResponse": {
        "type": "object",
        "properties": {
          "cardNumberOut": {
            "type": "string",
            "description": "Номер карты отправителя"
          },
          "balanceOut": {
            "type": "number",
            "description": "Баланс карты отправителя"
          },
          "cardNumberIn": {
            "type": "string",
            "description": "Номер карты получателя"
          },
          "balanceIn": {
            "type": "number",
            "description": "Баланс карты получателя"
          }
        },
        "description": "Возвращаемый объект при переводе средств между карт"
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PagedModelCardResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardResponse"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
      "PagedModelUserResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}